extract classes to test_helpers dir and add tests
edit tests and files in test_helpers dir

test_helpers/receiver.h fix
